pyproject.toml
[project]
name = "we-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "ronitrai27",email = "ronitrai1237@gmail.com"}
]
requires-python = ">=3.12,<4.0"
packages = [
    { include = "app" },
]
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn[standard] (>=0.44.0,<0.45.0)",
    "pydantic (>=2.13.2,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-core (>=1.3.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.1.14,<2.0.0)",
    "langgraph (>=1.1.8,<2.0.0)",
    "langgraph-checkpoint-postgres (>=3.0.5,<4.0.0)",
    "langsmith (>=0.7.32,<0.8.0)",
    "mem0ai (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "tenacity (>=9.1.4,<10.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "sse-starlette (>=2.1.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.11,<0.16.0)",
    "mypy (>=1.20.1,<2.0.0)"
]
